Correct: Adapter클래스 위치 변경
